CUI-322  Slider accessibility, fix pointer events and tooltip visibility
Depending on the z-order of the handles, Slider handles were not
showing the correct cursor on hover. Adding pointer-events: none to the
value divs, which wrap the handles, and pointer-events: auto to the
handles corrects this.

Tooltips should have visibility hidden as well as opacity: 0 so that
they do not respond to mouse events when they are visually hidden.