CUI-322 Fix for hidden tooltip while dragging handle in IE9
IE9 can't have a z-index or opacity on the handle which has a filter
applied or the tooltip will not render properly.

The following style from source/less/components/list.less adds both
z-index and opacity:

.dragging {
    .box-shadow(0 0.25rem 0.25rem rgba(0, 0, 0, 0.25));
    background-color: #eeeeee;
    opacity: 0.8;
    z-index: 10000;
}

This fix corrects the problem so that the tooltip remains visible when
dragging the slider in IE9.

Also improves behavior of pointer-events on tooltip so that the tooltip
is included in the handle's hit-area.