CUI-543 Improve accessibility of tri-state checkboxes.
Screen reader/browser combinations currently report checkboxes as
checked or unchecked but not "partially checked," even when we use
aria-checked="mixed"
and the indeterminate property. To work around this shortcoming, we
explicitly identify tri-state checkboxes with the css class "tristate,"
and replace the input[type=checkbox] elements with input[type=image]
with the appropriate WAI-ARIA role and state properties.

Conflicts:
	source/less/components/checkbox.less