CUI-1032 - Adds CSS for alert banner on page card
The changes refer to CQ-1344 and CQ-1444

Reusing existing alert components since the design and intention of the
page card banners are rather similar. Copying design of project card
banderole instead of the version from CQ-1444 since it is (a) way easier to
implement (b) does not need extra markup, and (c) looks similar to the
project card banderole.