CUI-7345 [Accessibility] Icon alt text should not default to a non-localized icon name.
The original intent of using the icon name as alternative text for accessibility, was to localize the icon name to some meaningful string, so that if no aria-label were provided to an icon-only button, the button would still be identified in a meaningful way to assistive technology. The current implementation simply uses the non-localized name for the Icon, which populates the interface with strings like ''Chevron Right" or "Chevron Down", which aren't helpful, and are problematic without proper translation.

It is better to let icon-only buttons without an aria-label, remain unlabelled so that automated testing tools have an opportunity to identify the missing label.