CUI-7343 [Accessibility] Coral.CycleButton dropdown should behave as a single menu
Per: https://jira.corp.adobe.com/browse/GRANITE-28179

Coral.CycleButton currently opens a dropdown containing a listbox of coral-cyclebutton-items, and a coral-buttonlist of coral-cyclebutton-actions, which makes interaction difficult for a keyboard or screen reader user, because it is not clear that the dropdown, that looks like a menu, contains more than one tab stop.

The dropdown should behave as a single menu with coral-cyclebutton-items having role=menuitemradio and coral-cyclebutton-actions having role=menuitem.