[Accessibility] Lists should not stopPropagation with keyboard navigation (#15)
In SelectList and ButtonList, keyboard event handlers for navigation with Arrow keys, PageUp/Down, Home or End should not call stopPropagation on the event.

In CycleButton, it is important for the CycleButton component to handle keyboard events from the SelectList and ButtonList in its Popover, so that the dropdown navigates as a single menu, rather than a dialog with two distinct tab stops that looks just like a menu.