IE11 Polyfill for coral-select-list-item querySelector on DOM-less elements
In IE11, when attempting to add a `coral-select-list-item` to a `coral-select-list`, if the `coral-select-list-item` is currently not within the DOM, then IE11 throws a SyntaxError on `content.querySelector('coral-icon:not(._coral-Menu-item-icon)')`. Specifically, it does not like the `:not` query selector for DOM-less elements, see https://github.com/FortAwesome/Font-Awesome/issues/12471.

This PR creates a polyfill for IE11, that ensures we can still remove the icons except `._coral-Menu-item-icon`s from the list item, without incurring a SyntaxError from IE11. This polyfill could have been extended to also check if the `content` is within the DOM, but I strongly suspect that tree-walking the DOM to find this element would be orders of magnitude slower than the array operations it does do.