CUI-7345 [Accessibility] Icon alt text should not default to a non-localized icon name.
The original intent of using the icon name as alternative text for accessibility, was to localize the icon name to some meaningful string, so that if no aria-label were provided to an icon-only button, the button would still be identified in a meaningful way to assistive technology. The current implementation simply uses the non-localized name for the Icon, which populates the interface with strings like ''Chevron Right" or "Chevron Down", which aren't helpful, and are problematic without proper translation.

It is better to let icon-only buttons without an aria-label, remain unlabelled so that automated testing tools have an opportunity to identify the missing label.

I've also laid the foundation for i18n by creating a translations.js file with variant-stripped icon names and suggested strings for things like Chevron Up/Down/Left/Right. An icon can be explicitly set to use the default string, which should be localized by setting `Coral.Icon.autoAriaLabel = true`.

This relates to CUI-4769 and CUI-4770