CUI-7391 Coral.Shell.MenuBar: NVDA workaround screen layout announcement of buttons
In browse mode, NVDA reads elements with display: inline-block as a single block of text. For the Shell.MenuBar, where buttons are rendered with display: inline-block, all the buttons are announced as a line, rather than one button being announced with each ArrowDown. This is the expected behavior for NVDA, and the user can exit NVDA's screen layout mode using NVDA+V keyboard shortcut. However, it may create some confusion.

A workaround may be to use display: flex on ._coral-Shell-menubar, so that the items still appear inline, but do not use display: inline-block on ._coral-Shell-menubar-item.