ColumnView: expanded/collapsed is announced as "row1 expanded" rather than cell name (#91)
With VoiceOver for macOS, ColumnView item name is not announced when the item receives focus, and the aria-expanded state change announces instead. This has to do with setting active and expanding the element on focus.

Attempted fix is to add setTimeout to use aria-hidden to hide the control when aria-expanded state changes, and then to remove aria-hidden after a delay. This delays the announcement of aria-expanded state, giving focused element enough time to be announced.