CUI-7402 ColumnView should support coral-interactive on item descendants
Like Accordion, some use cases of ColumnView require items to allow interactive descendants, like for example a checkbox.

To prevent keyboard and click events on the interactive descendants from bubbling up to be handled by the Item or ColumnView itself, we should support coral-interactive attribute to identify descendants that should be ignored.