CUI-7405 Coral.Shell.MenuBar.Item fix aria-expanded when set in response to coral-overlay:open or coral-overlay:close event
Per CQ-4293363#comment-21888821:

> 1. On pressing enter on any of the collapsed shell buttons like Solutions, Help, User, etc., NVDA briefly announces "expanded" but then it also announces "collapsed" again followed by "dialog" in some cases.

This seems to be caused in part by the rapid toggling of aria-expanded between coral-spectrum:beforeopen, coral-spectrum:open, and the setting of the open property, which was introduced with CUI-7390 and PR #88. We probably don't need to set aria-expanded on coral-spectrum:beforeopen, and should be careful to only update the attribute if its value is different from the current state of the overlay.