CUI-7416 Coral-Spectrum Tooltip is not included in accessibility name when referenced using aria-labelledby
In Chrome, foundation-autocomplete in AEM is not including the error tooltip text as part of the accessibility name even though the appropriate IDREF is included in the aria-labelledby attribute for the combobox input.

Possible solution is to set visibility: initial on the coral-tooltip animation to close has completed. Chrome seems to ignore the element with display: none, visibility: hidden, and aria-hidden="true", when calculating the accessibility name.