CUI-7422 wait a frame before updating attributes on selection
Per https://jira.corp.adobe.com/browse/CQ-4301452, the Metadata Schema editor relies on a check for `.is-selected` on selected items within an event handler for `coral-columnview:activeitemchange`. This works fine with CoralUI 3 version of component, that updates attributes within the sync function, but in Coral-Spectrum, we have to wait a frame, otherwise the current selection in the right hand column will be lost, before it is stored, when navigating to another active item in the left  hand column.