CUI-7422 ColumnView item should not be set active twice when using the mouse (#131)
* CUI-7422 ColumnView item should not be set active twice when using the mouse

Relating to https://jira.corp.adobe.com/browse/CQ-4301452, ColumnView items are being set to active twice. First with the focus event that comes on mouse down, and then again with the item click event on mouse up. This is causing choices to not be preserved in Granite's metadata schema editor, because when the item is already active, the item click event that follows does not trigger a `coral-columnview:activeitemchanged` event, which the schema editor uses to load the appropriate items and selection into the adjacent column: https://git.corp.adobe.com/Granite/com.adobe.granite.clientlibs.compiler/blob/master/src/test/resources/apps_libs5/libs/dam/gui/coral/components/admin/schemaforms/formbuilder/v2/clientlibs/js/cascadingschema/cascade.js#L176

* CUI-7422 wait a frame before updating attributes on selection

Per https://jira.corp.adobe.com/browse/CQ-4301452, the Metadata Schema editor relies on a check for `.is-selected` on selected items within an event handler for `coral-columnview:activeitemchange`. This works fine with CoralUI 3 version of component, that updates attributes within the sync function, but in Coral-Spectrum, we have to wait a frame, otherwise the current selection in the right hand column will be lost, before it is stored, when navigating to another active item in the left  hand column.

* CUI-7422 add code comments

Explain why item should be activated as well as focused if it is in a different column than the selected items.