fix(CUI-7441): Coral.CycleButton: fix menuitemradio and menuitem roles (#153)
Per https://jira.corp.adobe.com/browse/CUI-7441

In both the CoralUI and Coral Spectrum CycleButton implementations, the `menuitemradio` role on `coral-cyclebutton-item` and the `menuitem` role on `coral-cyclebutton-action` are not propagated to `coral-selectlist-item` and `button[is="coral-buttonlist-item"]` elements in the dropdown menu. Items within a menu should have a `menuitem*` role.

### Behavior can be seen in Coral Spectrum:
1. Open https://opensource.adobe.com/coral-spectrum/examples/#cyclebutton
2. Expand dropdown menu for the "With additional actions" CycleButton example.
3. Use the Web Inspector to inspect `coral-selectlist-item` elements in the `coral-selectlist[role="group"]` and `button[is="coral-buttonlist-item"]` in the `coral-buttonlist` contained within the `coral-popover[role="menu"]`.

#### Expected behavior
coral-selectlist-item elements should have role="menuitemradio"
button[is="coral-buttonlist-item"] elements should have role="menuitem"

#### Actual behavior
* `coral-selectlist-item` elements have `role="option"`, which is not appropriate for an item within `role="menu"`.
* `button[is="coral-buttonlist-item"]` elements have the implicit role of `button`, which is not appropriate for an item within `role="menu"`.

### Suggested fix
At CycleButton.js#L653 and CycleButton.js#L668, where we are defining items for the SelectList and actions for the ButtonList, we seem to expect the role from the `coral-cyclebutton-item` or `coral-cyclebutton-action` to get applied to the SelectList.Item or ButtonList.Item, but these components do not include getters and setters for the role attribute, so the role is not applied. We should instead either explicitly set the role attribute or add a getter/setter for role to SelectList.Item and ButtonList.Item.